다른 클라이언트에서 발사한 총알이 안보이던 문제 수정, 몬스터의 남은 hp 를 몬스터가 사망하는 시점에도 보내도록 수정

+++ OverflowJOLJAK/Shared/Protocol.h
@@ -40,7 +40,7 @@ struct sc_packet_monster_spawn : PACKET_HEADER      // 3 + 21바이트
     int m_id;
     unsigned char m_monster_type;
     float m_x, m_y, m_z;
-    int m_hp;
+    float m_hp;
 };
 
 struct sc_packet_monster_position : PACKET_HEADER   // 3 + 16바이트
@@ -65,7 +65,7 @@ struct cs_packet_player_attack : PACKET_HEADER      // 3 + 24바이트
 struct sc_packet_monster_hp : PACKET_HEADER         // 3 + 8바이트
 {
     int m_id;
-    int m_hp;
+    float m_hp;
 };
 
 struct sc_packet_monster_remove : PACKET_HEADER     // 3 + 4바이트
